Refactor DbContext (removed nested class)
